refactor: MealUiState와 ScheduleUiState를 별도의 파일에 선언
