+++ feature/main/src/main/java/com/practice/main/state/ScheduleUiState.kt
@@ -1,0 +1,26 @@
+package com.practice.main.state
+
+import com.practice.domain.schedule.Schedule
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.persistentListOf
+
+/**
+ * Ui state of schedule
+ */
+
+data class ScheduleUiState(
+    val schedules: ImmutableList<Schedule>,
+) {
+    val isEmpty: Boolean
+        get() = schedules.isEmpty()
+
+    val description: String
+        get() = if (schedules.isEmpty()) "학사일정이 없습니다." else schedules.joinToString(", ") { it.displayText }
+
+    companion object {
+        val EmptyScheduleState = ScheduleUiState(persistentListOf())
+    }
+}
+
+val Schedule.displayText: String
+    get() = if (eventName == eventContent) eventName else "$eventName - $eventContent"
